+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-17 20:04:49
-// Commit: test(perf): add unit tests for perf business logic
+// Updated at: 2026-04-18 09:40:54
+// Commit: feat(perf): implement responsive perf view component
 
 class perfComponentState {
-  final String commitRef = '494';
-  final String lastMessage = 'test(perf): add unit tests for perf business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-17T20:04:49+05:00');
+  final String commitRef = '506';
+  final String lastMessage = 'feat(perf): implement responsive perf view component';
+  final DateTime timestamp = DateTime.parse('2026-04-18T09:40:54+05:00');
 }