+++ CHANGELOG.md
@@ -507,3 +507,4 @@
 - [2026-04-17 23:13] refactor(state): decouple repository interface from concrete state service
 - [2026-04-17 23:29] refactor(cache): decouple repository interface from concrete cache service
 - [2026-04-18 09:32] test(theme): implement widget integration tests for theme
+- [2026-04-18 09:40] feat(perf): implement responsive perf view component